#6: make generic pipeline for 224,640 resolutions
