--- models/experimental/functional_yolov11/readme.md
@@ -1,11 +1,0 @@
-# Current status
-
-Facing ```ModuleNotFoundError: No module named 'ultralytics'``` issue while loading the weights file, so implemented with random weights
-- Finished reference implementation.
-- Completed ttnn implementation of Conv, Bottleneck, SPPF, C3K, C3K2
-
-## PCC:
-Conv - 0.99 </br>
-C3K2 with bottleneck - 0.99</br>
-C3K2 with C3k - facing shape mismatch error</br>
-SPPF - Not checked yet</br>
