--- tests/ttnn/integration_tests/yolov11/test_ttnn_sppf.py
@@ -1,54 +1,0 @@
-# SPDX-FileCopyrightText: © 2025 Tenstorrent Inc.
-# SPDX-License-Identifier: Apache-2.0
-
-import pytest
-import ttnn
-import torch
-from tests.ttnn.utils_for_testing import assert_with_pcc
-from models.experimental.functional_yolov11.tt.model_preprocessing import (
-    create_yolov11_input_tensors,
-    create_yolov11_model_parameters,
-)
-from models.experimental.functional_yolov11.reference.yolov11 import SPPF as torch_sppf
-from models.experimental.functional_yolov11.tt.ttnn_yolov11 import SPPF as ttnn_sppf
-
-
-@pytest.mark.parametrize(
-    "in_channel, out_channel, kernel, stride, padding, dilation, groups,fwd_input_shape",
-    [
-        ([256, 512], [128, 256], [1, 1], [1, 1], [0, 0], [1, 1], [1, 1], [1, 256, 7, 7]),
-    ],
-)
-@pytest.mark.parametrize("device_params", [{"l1_small_size": 79104}], indirect=True)
-def test_yolo_v11_sppf(
-    device,
-    use_program_cache,
-    reset_seeds,
-    in_channel,
-    out_channel,
-    kernel,
-    stride,
-    padding,
-    dilation,
-    groups,
-    fwd_input_shape,
-):
-    torch_module = torch_sppf(in_channel, out_channel, kernel, stride, padding, dilation, groups)
-    torch_module.eval()
-    torch_input, ttnn_input = create_yolov11_input_tensors(
-        device,
-        batch=fwd_input_shape[0],
-        input_channels=fwd_input_shape[1],
-        input_height=fwd_input_shape[2],
-        input_width=fwd_input_shape[3],
-    )
-    ttnn_input = ttnn.to_device(ttnn_input, device=device)
-    ttnn_input = ttnn.to_layout(ttnn_input, layout=ttnn.TILE_LAYOUT, memory_config=ttnn.L1_MEMORY_CONFIG)
-    torch_output = torch_module(torch_input)
-    parameters = create_yolov11_model_parameters(torch_module, torch_input, device=device)
-    ttnn_module = ttnn_sppf(device=device, parameter=parameters.conv_args, conv_pt=parameters)
-    ttnn_output = ttnn_module(x=ttnn_input, device=device)
-    ttnn_output = ttnn.to_torch(ttnn_output)
-    ttnn_output = ttnn_output.permute(0, 3, 1, 2)
-    ttnn_output = ttnn_output.reshape(torch_output.shape)
-    assert_with_pcc(torch_output, ttnn_output, 0.99999)
